homeフォルダを消して，homepageフォルダをhomeフォルダに変更
