+++ frontend/src/app/home/components/CourseList.tsx
