+++ frontend/src/app/home/components/EmptyCourse.tsx
